[7.x] Add `WithFixtures` to automatically loads fixtures file for test
