+++ src/Concerns/InteractsWithTestCase.php
@@ -179,6 +179,10 @@ trait InteractsWithTestCase
      */
     public static function setUpBeforeClassUsingTestCase(): void
     {
+        if (static::usesTestingConcern(WithFixtures::class)) {
+            static::setupWithFixturesForTestingEnvironment();
+        }
+
         static::resolvePhpUnitAttributesForMethod(static::class)
             ->flatten()
             ->filter(static fn ($instance) => $instance instanceof BeforeAllContract)
